fix: adds fallback when wasm is not supported
